+++ core/src/main/java/entidades/personajes/AtributosPersonaje/BarraCarga.java
@@ -1,4 +1,4 @@
-package entidades.personajes;
+package entidades.personajes.AtributosPersonaje;
 
 import com.badlogic.gdx.graphics.Color;
 import com.badlogic.gdx.graphics.glutils.ShapeRenderer;