+++ core/src/main/java/entidades/personajes/AtributosPersonaje/Mirilla.java
@@ -1,16 +1,13 @@
-package entidades.personajes;
+package entidades.personajes.AtributosPersonaje;
 
 import com.badlogic.gdx.graphics.Texture;
 import com.badlogic.gdx.graphics.g2d.Sprite;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
+import entidades.personajes.Personaje;
 import managers.GestorAssets;
 import utils.Constantes;
 
 public class Mirilla {
-
-    //Que la mirilla se muestre siempre que el personaje no se este moviendo
-    //Que su angulo se guarde, asi cuando se mueve, se oculta, y cuando se queda quieto vuelve aparecer
-    //En el ultimo angulo que estuvo
 
     private float x, y;
     private float angulo;
@@ -68,4 +65,3 @@ public class Mirilla {
     public void ocultarMirilla() { this.visible = false; }
     public float getAnguloRad() { return this.anguloRad; }
 }
-