+++ core/src/main/java/entidades/personajes/tiposPersonajes/HormigaExploradora.java
@@ -1,4 +1,4 @@
-package entidades.personajes;
+package entidades.personajes.tiposPersonajes;
 
 import Gameplay.Gestores.GestorColisiones;
 import Gameplay.Gestores.GestorProyectiles;
@@ -7,13 +7,14 @@ import Gameplay.Movimientos.Rango.LanzaGranada;
 import Gameplay.Movimientos.Rango.LanzaNectar;
 import Gameplay.Movimientos.Rango.LanzaRoca;
 import com.badlogic.gdx.graphics.Texture;
+import entidades.personajes.Personaje;
 import managers.GestorAssets;
 import utils.Constantes;
 
 public class HormigaExploradora extends Personaje {
 
     public HormigaExploradora(GestorColisiones gestorColisiones, GestorProyectiles gestorProyectiles, float x, float y) {
-        super(GestorAssets.get(Constantes.HORMIGA_EXPLORADORA, Texture.class), gestorColisiones, gestorProyectiles, x, y, 9999, 80, 150f);
+        super(GestorAssets.get(Constantes.HORMIGA_EXPLORADORA, Texture.class), gestorColisiones, gestorProyectiles, x, y, 9999, 150f);
     }
 
     @Override