+++ core/src/main/java/entidades/proyectiles/ProyectilesBalisticos/Roca.java
@@ -9,8 +9,10 @@ import utils.Constantes;
 
 public class Roca extends ProyectilExplosivo {
 
-    public Roca(float x, float y, float angulo, float velocidadBase, GestorColisiones gestorColisiones, Personaje ejecutor) {
-        super(x, y, angulo, velocidadBase, 10, gestorColisiones, ejecutor, 80, 140,
-            GestorAssets.get(Constantes.ROCA, Texture.class));
+    public Roca(float x, float y, float angulo, float velocidadBase,
+                GestorColisiones gestorColisiones, Personaje ejecutor) {
+        super(x, y, angulo, velocidadBase, 10, 450f, gestorColisiones, ejecutor,
+            80, 140, GestorAssets.get(Constantes.ROCA, Texture.class));
     }
 }
+