+++ core/src/main/java/utils/Constantes.java
@@ -5,8 +5,8 @@ public class Constantes {
     public static final int RESOLUCION_ANCHO = 1280;
     public static final int RESOLUCION_ALTO = 720;
 
-    public static final int RESOLUCION_ANCHO_MAPA = 2500;
-    public static final int RESOLUCION_ALTO_MAPA = 800;
+    public static final int RESOLUCION_ANCHO_MAPA = 3000;
+    public static final int RESOLUCION_ALTO_MAPA = 1300;
 
     public static final float GRAVEDAD = -980f;
     public static final float VEL_MAX_CAIDA = -2000f;